Move checkpoint parametrization out of common pkg
