+++ pkg/checkpoint/factory.go
@@ -1,7 +1,6 @@
 package checkpoint
 
 import (
-	"github.com/filecoin-project/mir/pkg/checkpoint/common"
 	"github.com/filecoin-project/mir/pkg/factorymodule"
 	"github.com/filecoin-project/mir/pkg/logging"
 	"github.com/filecoin-project/mir/pkg/modules"
@@ -9,7 +8,7 @@ import (
 	t "github.com/filecoin-project/mir/pkg/types"
 )
 
-func Factory(mc *common.ModuleConfig, ownID t.NodeID, logger logging.Logger) modules.PassiveModule {
+func Factory(mc *ModuleConfig, ownID t.NodeID, logger logging.Logger) modules.PassiveModule {
 	if logger == nil {
 		logger = logging.ConsoleErrorLogger
 	}
@@ -28,7 +27,7 @@ func Factory(mc *common.ModuleConfig, ownID t.NodeID, logger logging.Logger) mod
 				// Get the instance parameters
 				p := params.Type.(*factorypbtypes.GeneratorParams_Checkpoint).Checkpoint
 
-				chkpParams := &common.ModuleParams{
+				chkpParams := &ModuleParams{
 					OwnID:            ownID,
 					Membership:       p.Membership,
 					EpochConfig:      p.EpochConfig,
