+++ pkg/trantor/system.go
@@ -5,8 +5,6 @@ import (
 
 	"github.com/filecoin-project/mir/pkg/trantor/appmodule"
 
-	"github.com/filecoin-project/mir/pkg/checkpoint/common"
-
 	"github.com/pkg/errors"
 
 	"github.com/filecoin-project/mir/pkg/availability/batchdb/fakebatchdb"
@@ -127,7 +125,7 @@ func New(
 	}
 
 	// Factory module with instances of the checkpointing protocol.
-	checkpointing := checkpoint.Factory(common.DefaultModuleConfig(), ownID, logging.Decorate(logger, "CHKP: "))
+	checkpointing := checkpoint.Factory(checkpoint.DefaultModuleConfig(), ownID, logging.Decorate(logger, "CHKP: "))
 
 	// PBFT module with instances of the pbft protocol as segments to be called by ISS.
 	ordering := orderers.Factory(orderers.DefaultModuleConfig(), params.Iss, ownID, hashImpl, cryptoImpl, logging.Decorate(logger, "PBFT: "))
